+++ modules/apim-adapter/src/main/java/com/axway/apim/api/definition/WSDLSpecification.java
@@ -1,5 +1,9 @@
 package com.axway.apim.api.definition;
 
+import java.util.LinkedHashMap;
+
+import com.axway.apim.api.API;
+import com.axway.apim.api.model.ServiceProfile;
 import com.axway.apim.lib.errorHandling.AppException;
 import com.axway.apim.lib.utils.Utils;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -16,10 +20,24 @@ public class WSDLSpecification extends APISpecification {
 	public APISpecType getAPIDefinitionType() throws AppException {
 		return APISpecType.WSDL_API;
 	}
-
+	
 	@Override
 	public void configureBasepath(String backendBasepath) throws AppException {
-		
+		return;
+	}
+
+	@Override
+	public void configureBasepath(String backendBasepath, API api) throws AppException {
+		// For SOAP services, the WSDL has not been adapted in any case, 
+		// so the necessary service profile is to be created here.
+		if(backendBasepath!=null) {
+			ServiceProfile serviceProfile = new ServiceProfile();
+			serviceProfile.setBasePath(backendBasepath);
+			if(api.getServiceProfiles() == null) {
+				api.setServiceProfiles(new LinkedHashMap<String, ServiceProfile>());
+			}
+			api.getServiceProfiles().put("_default", serviceProfile);
+		}
 	}